requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
